+++ src/main/java/racingcar/controller/RacingGameController.java
@@ -3,25 +3,24 @@ package racingcar.controller;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import racingcar.PlayRequestDto;
-import racingcar.service.RacingGameService;
+import racingcar.dto.GameResultDto;
+import racingcar.dto.PlayRequestDto;
+import racingcar.service.RacingGameServiceImpl;
 
 import java.util.Arrays;
 import java.util.List;
 
-import static racingcar.service.RacingGameService.GameResult;
-
 @RestController
 public class RacingGameController {
 
-    private final RacingGameService racingGameService;
+    private final RacingGameServiceImpl racingGameService;
 
-    public RacingGameController(final RacingGameService racingGameService) {
+    public RacingGameController(final RacingGameServiceImpl racingGameService) {
         this.racingGameService = racingGameService;
     }
 
     @PostMapping("/plays")
-    GameResult play(
+    GameResultDto play(
             @RequestBody PlayRequestDto playRequestDto
     ) {
         List<String> names = Arrays.asList(playRequestDto.getNames().split(","));
