+++ src/test/java/racingcar/controller/RacingGameControllerTest.java
@@ -1,21 +1,23 @@
 package racingcar.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.hamcrest.Matchers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
-import racingcar.PlayRequestDto;
 import racingcar.domain.Car;
 import racingcar.domain.Cars;
 import racingcar.domain.Winners;
-import racingcar.service.RacingGameService;
+import racingcar.dto.GameResultDto;
+import racingcar.dto.PlayRequestDto;
+import racingcar.service.RacingGameServiceImpl;
 
 import java.util.List;
 
+import static org.hamcrest.Matchers.containsInAnyOrder;
+import static org.hamcrest.Matchers.is;
 import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.BDDMockito.given;
@@ -36,14 +38,14 @@ class RacingGameControllerTest {
     private ObjectMapper objectMapper;
 
     @MockBean
-    private RacingGameService racingGameService;
+    private RacingGameServiceImpl racingGameService;
 
     @Test
     void playTest() throws Exception {
         final String winnersPath = "$.winners";
-        final String carsPath = "$.racingCars[0:].position";
+        final String carsPath = "$..position";
         List<Car> cars = List.of(new Car("브리", 10), new Car("토미", 9));
-        RacingGameService.GameResult gameResult = new RacingGameService.GameResult(cars, new Winners(new Cars(cars)));
+        GameResultDto gameResult = new GameResultDto(cars, new Winners(new Cars(cars)));
         given(racingGameService.play(anyList(), anyInt()))
                 .willReturn(gameResult);
 
@@ -53,8 +55,8 @@ class RacingGameControllerTest {
                         .content(requestJson)
                         .contentType(APPLICATION_JSON))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath(winnersPath).value("브리"))
-                .andExpect(jsonPath(carsPath, Matchers.containsInAnyOrder(9, 10)))
+                .andExpect(jsonPath(winnersPath, is("브리")))
+                .andExpect(jsonPath(carsPath, containsInAnyOrder(9, 10)))
                 .andDo(print());
     }
 }
