+++ src/test/java/racingcar/service/RacingGameServiceImplTest.java
@@ -7,7 +7,7 @@ import org.springframework.boot.test.context.TestComponent;
 import org.springframework.context.annotation.Import;
 import org.springframework.transaction.annotation.Transactional;
 import racingcar.domain.numbergenerator.NumberGenerator;
-import racingcar.service.RacingGameService.GameResult;
+import racingcar.dto.GameResultDto;
 
 import javax.annotation.Priority;
 import java.util.List;
@@ -15,12 +15,12 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
-@Import(value = RacingGameServiceTest.MockNumberGenerator.class)
+@Import(value = RacingGameServiceImplTest.MockNumberGenerator.class)
 @Transactional
-class RacingGameServiceTest {
+class RacingGameServiceImplTest {
 
     @Autowired
-    private RacingGameService racingGameService;
+    private RacingGameServiceImpl racingGameService;
 
     @Priority(1)
     @TestComponent
@@ -39,7 +39,7 @@ class RacingGameServiceTest {
 
     @Test
     void playGame() {
-        GameResult gameResult = racingGameService.play(List.of("브리", "토미", "브라운"), 10);
+        GameResultDto gameResult = racingGameService.play(List.of("브리", "토미", "브라운"), 10);
 
         assertThat(gameResult.getWinners()).isEqualTo("브리");
         assertThat(gameResult.getRacingCars().size()).isEqualTo(3);
